Add config to disable cache for inventory

Approved-by: Rober Diaz
Approved-by: Pond

+++ app/Observers/SettingsObserver.php
@@ -37,8 +37,6 @@ class SettingsObserver
      */
     public function created(Settings $settings)
     {
-        $settings->load('website');
-
         $this->deleted($settings);
     }
 
@@ -50,8 +48,6 @@ class SettingsObserver
      */
     public function updated(Settings $settings)
     {
-        $settings->load('website');
-
         $this->deleted($settings);
     }
 
@@ -63,12 +59,16 @@ class SettingsObserver
      */
     public function deleted(Settings $settings)
     {
-        $website = $settings->website;
+        $settings->load('website');
 
-        $this->responseCache->forget(
-            $this->cacheKey->deleteByDealer($website->dealer_id),
-            $this->cacheKey->deleteSingleByDealer($website->dealer_id)
-        );
+        if (config('cache.inventory')) {
+            $website = $settings->website;
+
+            $this->responseCache->forget(
+                $this->cacheKey->deleteByDealer($website->dealer_id),
+                $this->cacheKey->deleteSingleByDealer($website->dealer_id)
+            );
+        }
 
         if ($dealerId = $settings->website->dealer_id) {
             dispatch(new PaymentCalculatorReIndexJob([$dealerId]));
